Split DeveloperScreen into a developer package
DeveloperScreen.kt was 1619 lines holding the ViewModel, the screen, ten
distinct cards, two stats tables, a paged card table and seven previews.
Continues the ui/word/ shape applied to Favorites, Settings and Search.

  ui/developer/DeveloperViewModel.kt     models, constants, ViewModel 522
  ui/developer/DeveloperScreen.kt        OptionsCard, Screen, Content  350
  ui/developer/DeveloperCards.kt         the seven action cards        469
  ui/developer/DeveloperStatsTables.kt   scheduling and family tables  275
  ui/developer/DeveloperCardTable.kt     paged card table              277
  ui/developer/preview/DeveloperPreviews.kt  the seven previews        244

DeveloperTerminalOverlay.kt moves in too: it is developer-only and only
App.kt mounts it.

DeveloperFormatters.kt deliberately stays in ui/. It reads as
developer-owned, but buildDeveloperDiagnosticInfo is used by
FavoriteSenseCard, so it is shared in fact if not in name. developer/
imports from it, as does favorites/.

Pure move. Diffing the moved bodies against the originals shows the
overlay byte-identical and the screen differing only by thirteen
widenings, each forced by a symbol now crossing a file boundary:

  TimeMachineCard, DeveloperNonDebugWarningCard, DeveloperTerminalCard,
  CompletionGalleryCard, IntakeCard, DeveloperCachesCard,
  DeveloperDangerCard, SchedulingStatsCard, CardTableCard,
  intakeModeLabel, TimeShiftOptions, CardTableColumnWidths,
  CardTableTotalWidth, CardTablePageSize   private -> internal

ScaryFilledTonalButton, the table row composables, CardTablePager,
cardTableHeaders and LoadingStatusRow keep private: each is used only
within the file it landed in.

ui/ drops from 25 files to 22.
