+++ composeApp/src/commonMain/kotlin/com/slovy/slovymovyapp/App.kt
@@ -40,6 +40,7 @@ import com.slovy.slovymovyapp.ui.*
 import com.slovy.slovymovyapp.ui.favorites.*
 import com.slovy.slovymovyapp.ui.settings.*
 import com.slovy.slovymovyapp.ui.search.*
+import com.slovy.slovymovyapp.ui.developer.*
 import com.slovy.slovymovyapp.data.lists.ListsService
 import com.slovy.slovymovyapp.data.lists.WordListsRepository
 import com.slovy.slovymovyapp.data.recovery.RecoverableSense
