+++ composeApp/src/commonMain/kotlin/com/slovy/slovymovyapp/ui/developer/DeveloperTerminalOverlay.kt
@@ -1,4 +1,9 @@
-package com.slovy.slovymovyapp.ui
+package com.slovy.slovymovyapp.ui.developer
+
+import com.slovy.slovymovyapp.ui.ThemePreviewProvider
+import com.slovy.slovymovyapp.ui.ThemedPreview
+import com.slovy.slovymovyapp.ui.developerLogSignature
+import com.slovy.slovymovyapp.ui.toDeveloperTerminalLine
 
 import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.clickable
@@ -36,11 +41,9 @@ import androidx.compose.material3.TextButton
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.pointer.pointerInput
@@ -67,6 +70,8 @@ import kotlin.time.Duration.Companion.days
 import kotlin.time.Duration.Companion.hours
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.Duration.Companion.minutes
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
 
 private const val TAG = "DeveloperTerminalOverlay"
